refactor ClubVC for MVP
